# shared/openapi/admin.yaml
openapi: 3.0.0
info:
  title: Samaagum Admin Unified API
  version: 1.0.0
  description: Type-safe Unified API contract for all Samaagum Admin Console services (RBAC, plans, coupons, users, KYC, disputes, moderation, tenants, feature flags, and audit logs).
servers:
  - url: http://localhost:3000
    description: Local development backend server
paths:
...
  /api/admin/plans:
    get:
      summary: Retrieve all plans
      operationId: getPlans
      responses:
        '200':
          description: A list of plans
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      plans:
                        type: array
                        items:
                          $ref: '#/components/schemas/Plan'
    post:
      summary: Create a plan
      operationId: createPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanInput'
      responses:
        '201':
          description: Plan created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /api/admin/plans/{id}:
    put:
      summary: Update a plan
      operationId: updatePlan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanInput'
      responses:
        '200':
          description: Plan updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
    delete:
      summary: Delete a plan
      operationId: deletePlan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plan deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  # ── Coupons paths ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Plan:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        category:
          type: string
        plan_type:
          type: string
        is_active:
          type: boolean
        is_popular:
          type: boolean
        group_name:
          type: string
        pricing:
          type: object
        metadata:
          type: object
        trial:
          type: object
        visibility:
          type: object
        rbac_role_id:
          type: string
        rbac_position_id:
          type: string
        rbac_auto_assign:
          type: boolean
    PlanInput:
      type: object
      required:
        - name
        - displayName
      properties:
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        category:
          type: string
        planType:
          type: string
        isActive:
          type: boolean
        isPopular:
          type: boolean
        groupName:
          type: string
        pricing:
          type: object
        metadata:
          type: object
        trial:
          type: object
        visibility:
          type: object
        rbac:
          type: object
...